address comments from ashley
